fixed error with overflow of answer

+++ pizza_cutting_10079.c
@@ -4,14 +4,14 @@
 int main()
 {
     int n;
-    while(scanf("%d\n", &n) != EOF && n >= 0) {
+    while(scanf("%u\n", &n) != EOF && n >= 0) {
         int i;
-        int slices = 1;
+        long long unsigned int slices = 1;
         for(i = 1; i <= n; i++) {
             slices += i;
 /* printf("%d %d\n", i, slices);             */
         }
-        printf("%d\n", slices);            
+        printf("%llu\n", slices);            
     }
 
     return 0;
